adding all vehicles and weapons to a one file respectively and removing the packages
