--- katas/metal-slug/kevin/vehicles/Camel.js
@@ -1,7 +1,0 @@
-import { Vehicle } from './Vehicle.js';
-export class Camel extends Vehicle {
-    constructor(lifePoints) {
-        super(lifePoints);
-        this.setvehicleLifePoints(1);
-    }
-}
