--- katas/metal-slug/kevin/vehicles/Plane.js
@@ -1,7 +1,0 @@
-import { Vehicle } from './Vehicle.js';
-export class Plane extends Vehicle {
-    constructor(lifePoints) {
-        super(lifePoints);
-        this.setvehicleLifePoints(4);
-    }
-}
