--- katas/metal-slug/kevin/vehicles/Submarine.js
@@ -1,7 +1,0 @@
-import { Vehicle } from './Vehicle.js';
-export class Submarine extends Vehicle {
-    constructor(lifePoints) {
-        super(lifePoints);
-        this.setvehicleLifePoints(3);
-    }
-}
