--- katas/metal-slug/kevin/vehicles/Tank.js
@@ -1,7 +1,0 @@
-import { Vehicle } from './Vehicle.js';
-export class Tank extends Vehicle {
-    constructor(lifePoints) {
-        super(lifePoints);
-        this.setvehicleLifePoints(3);
-    }
-}
