--- katas/metal-slug/kevin/weapons/FlameShot.js
@@ -1,7 +1,0 @@
-import { Weapon } from './Weapon.js';
-export class FlameShot extends Weapon {
-    constructor(attackPower) {
-        super(attackPower);
-        this.setweaponAttackPower(5);
-    }
-}
