--- katas/metal-slug/kevin/weapons/HeavyMachineGun.js
@@ -1,7 +1,0 @@
-import { Weapon } from './Weapon.js';
-export class HeavyMachineGun extends Weapon {
-    constructor(attackPower) {
-        super(attackPower);
-        this.setweaponAttackPower(2);
-    }
-}
