--- katas/metal-slug/kevin/weapons/Shotgun.js
@@ -1,7 +1,0 @@
-import { Weapon } from './Weapon.js';
-export class Shotgun extends Weapon {
-    constructor(attackPower) {
-        super(attackPower);
-        this.setweaponAttackPower(5);
-    }
-}
