--- katas/metal-slug/kevin/weapons/Weapon.js
@@ -1,11 +1,0 @@
-export class Weapon {
-    constructor(attackPower) {
-        this.attackPower = attackPower;
-    }
-    getweaponAttackPower() {
-        return this.attackPower;
-    }
-    setweaponAttackPower(attackPower) {
-        this.attackPower = attackPower;
-    }
-}
